Refactor: Move MatchUiState to mappers package
The `MatchUiState` data class was moved from the `com.karrar.movieapp.ui.match` package to the `com.karrar.movieapp.ui.mappers` package.

The import statement in `MatchMapper.kt` was updated to reflect this change. Additionally, `MatchUiState.kt` now imports `ExploreScreenState`, `MatchPages`, and `MovieScreenState` from the `com.karrar.movieapp.ui.match` package.
